Updated backend :
Updated pom to latest version, switched back from NC to SC.

Removed unnecessary packages from NPM, updated day of week to include classes on previous and next button to add svg, will need to find a different svg as it's not directly in the middle so rotating it in css does not do it justice.

Updated the popup on the leaflet map to open seperately from the map.

Probably next step is to hide the filter and start work on full text search of the database.